backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
google-api-python-client
google-auth
